CovidAnalysis/FunctionalPrograming/requirements.txt
pandas
numpy
django
djangorestframework
requests